refactor(chat): extract objective system + realism eval plumbing
Two more cohesive clusters move out of the chat_service.dart coordinator into
`part`/extension files (verbatim, zero behaviour change):

- Objective System — load/inject/set/clear objectives, all task management
  (toggle/update/add/remove/complete), and background completion checks
  -> chat_service_objectives.dart
- Realism eval plumbing — chance-time injection, LLM eval fire + think-block
  stripping, the five eval-call wrappers, trust-repair, realism-state capture,
  and staggered eval dispatch
  -> chat_service_realism_evals.dart

Three members had to stay in the class body rather than move with their
clusters, each for a concrete reason that's now documented inline so a future
extractor doesn't repeat the work:

- getActiveObjectivesFor and isCheckingCompletion are overridden by
  FakeChatService in the golden tests. Extension members are statically
  dispatched, so an `@override` on them doesn't bind — they must remain part of
  the class interface.
- _kEvalDispatchStagger is a static const referenced by the periodic-eval
  coordinator that's still in chat_service.dart. Extension statics aren't
  visible unqualified to the host type, so it stays on the class and the
  extension references it as ChatService._kEvalDispatchStagger.

Net effect: chat_service.dart drops 4,742 -> 4,213 lines (-529). No public API,
persistence shape, or realism/needs behaviour changes.

flutter analyze: clean. Full unit suite (1348) + golden suite (94) + the
1:1<->group realism parity gate: green. (A single BackgroundSettingsDialog
golden flakes only under combined unit+golden concurrency — unrelated to this
change; it passes in isolation and in the golden-only run.)
