+++ lib/services/chat_service.dart
@@ -97,6 +97,8 @@ part 'chat/chat_service_chat_entry.dart';
 part 'chat/chat_service_group_entry.dart';
 part 'chat/chat_service_session_state.dart';
 part 'chat/chat_service_session_load.dart';
+part 'chat/chat_service_realism_evals.dart';
+part 'chat/chat_service_objectives.dart';
 part 'chat/chat_service_realism_dance.dart';
 part 'chat/chat_service_speaker_objectives.dart';
 part 'chat/chat_service_impersonate.dart';
@@ -161,6 +163,13 @@ class ChatService extends ChangeNotifier {
       _activeObjectives.where((o) => o.isPrimary).firstOrNull;
   List<Objective> get secondaryObjectives =>
       _activeObjectives.where((o) => !o.isPrimary).toList();
+
+  /// Whether a completion check is currently running.
+  ///
+  /// Kept in the class body (not the objectives extension) because
+  /// [FakeChatService] overrides it in golden tests — extension members are
+  /// statically dispatched and cannot be overridden.
+  bool get isCheckingCompletion => _isCheckingCompletion;
 
   List<Map<String, dynamic>> tasksForObjective(Objective obj) {
     try {
@@ -2339,6 +2348,12 @@ class ChatService extends ChangeNotifier {
       '- Use *asterisks* for actions and narration: *She leaned against the doorframe, arms crossed.*\n'
       '- Internal thoughts can be written in italics or described through narration.';
 
+  /// Inter-call delay used when staggering the multi-call realism evaluations.
+  /// Kept in the class body (not the realism-evals extension) because the
+  /// periodic-eval coordinator in this file references it directly; extension
+  /// statics aren't visible unqualified to the host type.
+  static const _kEvalDispatchStagger = Duration(milliseconds: 50);
+
   CharacterCard? get activeCharacter => _activeCharacter;
   List<ChatMessage> get messages => List.unmodifiable(_messages);
   bool get isGenerating => _isGenerating;
@@ -3626,335 +3641,6 @@ class ChatService extends ChangeNotifier {
     }
   }
 
-  // ── Objective System ───────────────────────────────────────────────────
-
-  /// Load the active objectives for the current session from DB.
-  Future<void> _loadActiveObjectives() async {
-    if (_activeCharacter == null || _currentSessionId == null) {
-      _activeObjectives = [];
-      _messagesSinceLastCheck = 0;
-      _isCheckingCompletion = false;
-      _summaryPaused =
-          false; // explicit secondary zero for _summaryPaused (symmetric; _loadActiveObjectives empty hygiene)
-      _isSummaryGenerating =
-          false; // secondary zero in _loadActiveObjectives empty (0-session hygiene for summary flag)
-      _userMessagesSinceLastPeriodicEval = 0;
-      _isExtractingFacts =
-          false; // secondary fact flag + counter zero in _loadActiveObjectives empty (0-session hygiene; fact_extraction)
-      _isEvolvingCharacter = false;
-      _evolutionStatus = '';
-      _evolutionError =
-          ''; // explicit evo flag/status/error zero in _loadActiveObjectives empty (0-session hygiene; evolution_service (stateless or prompt-only; no reset calls needed))
-      return;
-    }
-    final charId = _getCharacterIdFromCard(_activeCharacter!);
-    try {
-      _activeObjectives = await _db.getActiveObjectives(
-        charId,
-        chatId: _currentSessionId!,
-      );
-      for (final obj in _activeObjectives) {
-        debugPrint(
-          '[Objective] Loaded: ${obj.objective} (Primary: ${obj.isPrimary})',
-        );
-      }
-    } catch (e) {
-      debugPrint(
-        '[Objective] Failed to load (will run without objectives this session): $e',
-      );
-      _activeObjectives = [];
-    }
-    notifyListeners(); // Central _disposed guard in ChatService overrides now protects this (and all other) post-async notify sites. Per-site try/catch removed (deletion part of rec 2 task); see god _disposed + notify override + setActiveCharacter:2205 comment.
-  }
-
-  /// Build the prompt injection text for the active objectives.
-  /// Wording intensity varies based on injection depth for the primary objective.
-  /// Secondary objectives are injected as ambient background goals.
-  String _getObjectiveInjection() {
-    // Thin delegation (full in AuthorNoteBuilder per step 8). Objective state mgmt
-    // (lists, getters, tasksFor) stays in god (objective_service is later step).
-    return _authorNoteBuilder.buildObjectiveInjection();
-  }
-
-  /// Set a new objective for the current session (or for a specific character when in group mode).
-  ///
-  /// [autoGenerateTasks] defaults to false. User-created objectives (typed in the UI) should
-  /// not auto-generate subtasks — the user is in control of their own quests and can use the
-  /// explicit "Generate Tasks" button if desired.
-  ///
-  /// Autonomous objectives proposed by the character (via the realism "proposed_objective"
-  /// evals) pass true so that the character's self-generated goals come with concrete
-  /// sequential tasks. This makes the AI-driven objectives feel organic and like something
-  /// the character is actively striving to accomplish.
-  Future<void> setObjective(
-    String goal, {
-    bool isPrimary = true,
-    CharacterCard? targetCharacter,
-    bool autoGenerateTasks = false,
-  }) async {
-    if (goal.trim().isEmpty) return;
-    if (_currentSessionId == null) return;
-
-    CharacterCard? target = targetCharacter;
-    if (target == null) {
-      if (_activeGroup != null) {
-        // During per-speaker group realism evals (which propose autonomous objectives),
-        // _activeCharacter is temporarily impersonated to the evaluated speaker. Prefer it
-        // so the character's own internal goal attaches to *them*, not nextCharacter.
-        final currentIsGroupMember =
-            _activeCharacter != null &&
-            _groupCharacters.any(
-              (c) =>
-                  _getCharacterIdFromCard(c) ==
-                  _getCharacterIdFromCard(_activeCharacter!),
-            );
-        if (currentIsGroupMember) {
-          target = _activeCharacter;
-        } else {
-          target = nextCharacter ?? _groupCharacters.firstOrNull;
-        }
-      } else {
-        target = _activeCharacter;
-      }
-    }
-    if (target == null) return;
-
-    final charId = _getCharacterIdFromCard(target);
-
-    if (isPrimary) {
-      final existing = await _db.getObjectivesForCharacter(
-        charId,
-        chatId: _currentSessionId,
-      );
-      for (final obj in existing) {
-        if (obj.active && obj.isPrimary) {
-          await _db.updateObjective(
-            ObjectivesCompanion(
-              id: drift.Value(obj.id),
-              isPrimary: const drift.Value(false),
-            ),
-          );
-        }
-      }
-    } else {
-      final currentSecondaries = secondaryObjectives;
-      if (currentSecondaries.length >= 2) {
-        for (int i = 0; i < currentSecondaries.length - 1; i++) {
-          await _db.updateObjective(
-            ObjectivesCompanion(
-              id: drift.Value(currentSecondaries[i].id),
-              active: const drift.Value(false),
-            ),
-          );
-        }
-      }
-    }
-
-    final newId = const Uuid().v4();
-    await _db.insertObjective(
-      ObjectivesCompanion.insert(
-        id: newId,
-        characterId: charId,
-        objective: goal.trim(),
-        chatId: drift.Value(_currentSessionId),
-        active: const drift.Value(true),
-        isPrimary: drift.Value(isPrimary),
-      ),
-    );
-
-    await _loadActiveObjectives();
-    _messagesSinceLastCheck = 0;
-
-    if (autoGenerateTasks) {
-      try {
-        final forChar = await getActiveObjectivesFor(target);
-        final matches = forChar.where((o) => o.id == newId);
-        final addedObj = matches.isNotEmpty ? matches.first : null;
-        if (addedObj != null) {
-          unawaited(
-            generateObjectiveTasks(
-              addedObj,
-              taskCount: 3,
-              nsfw: false,
-            ), // step 11 thin (full in objective_proposal)
-          );
-        }
-      } catch (_) {
-        // Objective created successfully; task generation is best-effort and non-fatal.
-        // User can always tap "Generate Tasks" manually.
-      }
-    }
-  }
-
-  /// Generate subtasks for the current objective using the LLM.
-  /// Clears existing tasks first so regen always produces a clean slate.
-  // Thin delegation (full generateObjectiveTasks + 2000 budget + central strip + proposal
-  // handling in objective_proposal step 11; objective mgmt coordination / list / load / db
-  // updates stayed thin in god per plan for step9/11; "thin delegation here; full objective
-  // proposal in step 11").
-  Future<void> generateObjectiveTasks(
-    Objective obj, {
-    int taskCount = 5,
-    bool nsfw = false,
-  }) => _objectiveProposal.generateObjectiveTasks(
-    obj,
-    taskCount: taskCount,
-    nsfw: nsfw,
-  );
-
-  /// Marks the first uncompleted task matching taskDesc as completed (best-effort side-effect
-  /// for auto-complete in checkTaskCompletionInBackground currentTask YES path).
-  /// (Thin delegation; full mutation logic here in god per plan for step 11 to keep list/db
-  /// mutation thin/stayed in god; leaf calls via cb. Matches toggleTask pattern exactly.)
-  Future<void> markTaskCompleted(Objective obj, String taskDesc) async {
-    final tasks = tasksForObjective(obj);
-    final idx = tasks.indexWhere(
-      (t) => (t['description'] as String) == taskDesc && t['completed'] != true,
-    );
-    if (idx < 0) return;
-    tasks[idx]['completed'] = true;
-    await _db.updateObjective(
-      ObjectivesCompanion(
-        id: drift.Value(obj.id),
-        tasks: drift.Value(jsonEncode(tasks)),
-      ),
-    );
-    await _loadActiveObjectives();
-  }
-
-  /// Manually toggle a task's completion status.
-  Future<void> toggleTask(Objective obj, int taskIndex) async {
-    final tasks = tasksForObjective(obj);
-    if (taskIndex < 0 || taskIndex >= tasks.length) return;
-
-    tasks[taskIndex]['completed'] = !(tasks[taskIndex]['completed'] as bool);
-    await _db.updateObjective(
-      ObjectivesCompanion(
-        id: drift.Value(obj.id),
-        tasks: drift.Value(jsonEncode(tasks)),
-      ),
-    );
-    await _loadActiveObjectives();
-  }
-
-  /// Update the description of a specific task.
-  Future<void> updateTask(
-    Objective obj,
-    int taskIndex,
-    String newDescription,
-  ) async {
-    final tasks = tasksForObjective(obj);
-    if (taskIndex < 0 || taskIndex >= tasks.length) return;
-    if (newDescription.trim().isEmpty) return;
-
-    tasks[taskIndex]['description'] = newDescription.trim();
-    await _db.updateObjective(
-      ObjectivesCompanion(
-        id: drift.Value(obj.id),
-        tasks: drift.Value(jsonEncode(tasks)),
-      ),
-    );
-    await _loadActiveObjectives();
-  }
-
-  /// Clear the active objective.
-  Future<void> clearObjective(Objective obj) async {
-    await _db.updateObjective(
-      ObjectivesCompanion(
-        id: drift.Value(obj.id),
-        active: const drift.Value(false),
-      ),
-    );
-    await _loadActiveObjectives();
-    _messagesSinceLastCheck = 0;
-  }
-
-  /// Update the injection depth for the active objective.
-  Future<void> updateObjectiveDepth(Objective obj, int depth) async {
-    await _db.updateObjective(
-      ObjectivesCompanion(
-        id: drift.Value(obj.id),
-        injectionDepth: drift.Value(depth),
-      ),
-    );
-    await _loadActiveObjectives();
-  }
-
-  /// Add a manually created task to the active objective.
-  Future<void> addManualTask(Objective obj, String description) async {
-    if (description.trim().isEmpty) return;
-    final tasks = tasksForObjective(obj);
-    tasks.add({'description': description.trim(), 'completed': false});
-    await _db.updateObjective(
-      ObjectivesCompanion(
-        id: drift.Value(obj.id),
-        tasks: drift.Value(jsonEncode(tasks)),
-      ),
-    );
-    await _loadActiveObjectives();
-  }
-
-  /// Remove a task from the active objective.
-  Future<void> removeTask(Objective obj, int taskIndex) async {
-    final tasks = tasksForObjective(obj);
-    if (taskIndex < 0 || taskIndex >= tasks.length) return;
-    tasks.removeAt(taskIndex);
-    await _db.updateObjective(
-      ObjectivesCompanion(
-        id: drift.Value(obj.id),
-        tasks: drift.Value(jsonEncode(tasks)),
-      ),
-    );
-    await _loadActiveObjectives();
-  }
-
-  /// Update how often task completion is checked.
-  Future<void> updateCheckFrequency(Objective obj, int frequency) async {
-    await _db.updateObjective(
-      ObjectivesCompanion(
-        id: drift.Value(obj.id),
-        checkFrequency: drift.Value(frequency),
-      ),
-    );
-    await _loadActiveObjectives();
-  }
-
-  /// Check if the current task has been completed (called periodically).
-  /// Manually trigger a completion check (called from UI "Check now" button).
-  void forceCheckCompletion() {
-    if (_activeObjectives.isEmpty) return;
-    _checkTaskCompletionInBackground(); // step 11 thin (full in objective_proposal)
-    notifyListeners(); // trigger UI to show spinner
-  }
-
-  /// Whether a completion check is currently running.
-  bool get isCheckingCompletion => _isCheckingCompletion;
-
-  /// Synchronous version — awaits the check. Used pre-generation.
-  Future<void> _maybeCheckTaskCompletionSync() async {
-    if (_activeObjectives.isEmpty ||
-        _llmProvider == null ||
-        _isCheckingCompletion) {
-      return;
-    }
-
-    _messagesSinceLastCheck++;
-    final freq = _realismEnabled
-        ? 1
-        : (primaryObjective?.checkFrequency ??
-              _activeObjectives.first.checkFrequency);
-    if (_messagesSinceLastCheck < freq) return;
-    _messagesSinceLastCheck = 0;
-
-    await _checkTaskCompletionInBackground(); // step 11 thin (full in objective_proposal)
-  }
-
-  // Thin delegation (full _checkTaskCompletionInBackground + 2000 budget + central strip in
-  // objective_proposal step 11; objective mgmt coordination / isChecking flag / load / db
-  // updates stayed thin in god per plan for step9/11; "thin delegation here; full objective
-  // proposal in step 11").
-  Future<void> _checkTaskCompletionInBackground() =>
-      _objectiveProposal.checkTaskCompletionInBackground();
 
   // Two god-owned "since last" counters for the independent periodic features.
   // Facts/auto-persona uses the first (tied to autoPersonaInterval).
@@ -4284,221 +3970,6 @@ class ChatService extends ChangeNotifier {
   // The sub-builders themselves are still instantiated and passed to the composer.
   // Chance Time remains separate (it is not part of the per-turn realism state bundle).
 
-  /// Injects a Chance Time event into the character's response prompt.
-  /// Placed AFTER the character name suffix for maximum recency weight.
-  /// Consumed after one use (cleared after response generation).
-  String _getChanceTimeInjection() {
-    // Thin delegation (full in ChaosInjection per step 8; UI flags stayed in god per plan).
-    return _chaosInjection.buildChanceTimeInjection();
-  }
-
-  // ── LLM Eval Thins (step 9; full in LlmEvalEngine) + Needs Impact Thins (consolidated) + Objective Proposal Thins (step 11) ──
-  // 0 new god privates beyond required thin delegates (fire/strip/extract/evaluate* thins + _runPostGenNeedsChecks thin (consolidated to evaluator; the prior separate _check* bodies excised as dead/vestigial per task) + generate/_check thins for objective; void_ count 15; +1 late final); thins only (public surface for now per plan); objective proposal coordination + some
-  // prompt/obj mgmt + post-gen needs orchestration (impersonation dance, pre/post group scalars, long-gen, metadata attach) stayed thin in god per plan (qualified in objective_proposal header + here + test + MD).
-  // All call sites (5 firing points for realism evals now via realism_evals step 10, gen/check now via objective_proposal step 11, proposal, direct fire/strip/extract in eval paths, post-gen needs) now delegate; non-eval uses ... also route via these thins (centralized, no parallel).
-
-  Future<String?> _fireLLMEval(
-    String prompt, {
-    void Function(String)? onChunk,
-  }) => _llmEvalEngine.fireLLMEval(prompt, onChunk: onChunk);
-
-  String _stripThinkBlocks(String text) =>
-      _llmEvalEngine.stripThinkBlocks(text);
-
-  int? _extractJsonInt(String text, String key) =>
-      _llmEvalEngine.extractJsonInt(text, key);
-
-  bool? _extractJsonBool(String text, String key) =>
-      _llmEvalEngine.extractJsonBool(text, key);
-
-  // KoboldCpp receives HTTP requests in wire order via loopback.
-  // A small stagger prevents TCP timing from reordering concurrent
-  // eval dispatches, ensuring KoboldCpp's FIFO queue (which serializes
-  // internally) processes evals in our intended order rather than
-  // reverse or interleaved. Zero wall time added — KoboldCpp serializes
-  // anyway, so the stagger just ensures already-in-flight ordering.
-  static const _kEvalDispatchStagger = Duration(milliseconds: 50);
-
-  Future<void> _evaluateRelationshipCall({void Function(String)? onChunk}) =>
-      _realismEvals.evaluateRelationshipCall(onChunk: onChunk);
-
-  Future<void> _evaluateEmotionalStateCall({void Function(String)? onChunk}) =>
-      _realismEvals.evaluateEmotionalStateCall(onChunk: onChunk);
-
-  Future<void> _evaluatePhysicalStateCall({void Function(String)? onChunk}) =>
-      _realismEvals.evaluatePhysicalStateCall(onChunk: onChunk);
-
-  Future<void> _evaluateNarrativeCall({void Function(String)? onChunk}) =>
-      _realismEvals.evaluateNarrativeCall(onChunk: onChunk);
-
-  Future<void> _evaluateOneShotCall({void Function(String)? onChunk}) =>
-      _realismEvals.evaluateOneShotCall(onChunk: onChunk);
-
-  /// One-shot trust repair evaluator.
-  ///
-  /// Called automatically on the user's next message after a severe trust drop
-  /// (≥ -20 delta). Replaces the normal relationship eval for that turn.
-  /// The LLM weighs the explanation against character persona and chat history,
-  /// returning a trust_recovery value (0–60). Recovery is capped to prevent
-  /// instant restoration from Absolute Distrust.
-  Future<void> _evaluateTrustRepairCall(
-    String userExplanation, {
-    void Function(String)? onChunk,
-  }) async {
-    if (!_realismEnabled || _activeCharacter == null) return;
-
-    if (_activeCharacter == null) {
-      // Group chat or other mode — relationship evals not supported in this path yet
-      return;
-    }
-    final charName = _activeCharacter!.name;
-    final persona = _activeCharacter!.personality;
-    final recentCount = _messages.length < 10 ? _messages.length : 10;
-    final history = _messages.reversed
-        .take(recentCount)
-        .toList()
-        .reversed
-        .map((m) => '${m.sender}: ${m.displayText}')
-        .join('\n');
-
-    final prompt =
-        'You are evaluating whether $charName should partially restore trust '
-        'after a severe breach caused by the previous interaction.\n\n'
-        'Character Persona: $persona\n\n'
-        'Recent chat history (last ~10 messages):\n$history\n\n'
-        'The user\'s trust-repair explanation is: "$userExplanation"\n\n'
-        'Evaluate ONLY whether this explanation is convincing given:\n'
-        '1. The character\'s personality — are they forgiving, stubborn, paranoid, naive?\n'
-        '2. The plausibility of the explanation against the chat history\n'
-        '3. Whether the explanation contradicts established facts\n\n'
-        'Rules:\n'
-        '- trust_recovery: 0 (rejected) to 60 (fully convincing)\n'
-        '- Paranoid/skeptical characters: give 0–20 even for good explanations\n'
-        '- Forgiving/naive characters: may give 30–60 for plausible explanations\n'
-        '- Do NOT give 60 unless the explanation perfectly resolves the breach\n'
-        '- "reason" must be 1 short sentence from the character\'s POV\n\n'
-        'Respond with ONLY: {"trust_recovery": <0-60>, "verdict": "accepted|partial|rejected", "reason": "<brief>"}\n';
-
-    try {
-      debugPrint('[Realism:TrustRepair] Evaluating repair attempt...');
-      final raw = await _fireLLMEval(prompt, onChunk: onChunk);
-      if (raw == null) return;
-
-      final text = _stripThinkBlocks(raw).trim();
-
-      final verdictMatch = RegExp(
-        r'"verdict"\s*:\s*"([^"]+)"',
-      ).firstMatch(text);
-      final reasonMatch = RegExp(r'"reason"\s*:\s*"([^"]*)"').firstMatch(text);
-
-      final recovery = (_extractJsonInt(text, 'trust_recovery') ?? 0).clamp(
-        0,
-        60,
-      );
-      final verdict = verdictMatch?.group(1) ?? 'rejected';
-      final reason = reasonMatch?.group(1) ?? '';
-
-      if (recovery > 0) {
-        _relationshipService.applyTrustDelta(recovery);
-        debugPrint(
-          '[Realism:TrustRepair] $verdict — recovered $recovery → ${_relationshipService.trustLevel} ($reason)',
-        );
-      } else {
-        debugPrint('[Realism:TrustRepair] Rejected — no recovery ($reason)');
-      }
-
-      // Surface verdict in message metadata so swipe history can record it
-      _pendingRealismMetadata = {
-        ...?_pendingRealismMetadata,
-        'trust_repair_verdict': verdict,
-        'trust_repair_recovery': recovery,
-        if (reason.isNotEmpty) 'trust_repair_reason': reason,
-      };
-
-      _saveChat();
-      notifyListeners();
-    } catch (e) {
-      debugPrint('[Realism:TrustRepair] Failed: $e');
-    }
-  }
-
-  Map<String, dynamic> _captureRealismState({Map<String, int>? preTurn}) {
-    final state = {
-      'affectionScore': _relationshipService.affectionScore,
-      'relationshipTier': _relationshipService.relationshipTier,
-      'longTermScore': _relationshipService.longTermScore,
-      'longTermTier': _relationshipService.longTermTier,
-      'turnsSinceLongTermCheck': _relationshipService.turnsSinceLongTermCheck,
-      'shortTermDeltasSummary': _relationshipService.shortTermDeltasSummary,
-      'moodDecayCounter': _moodDecayCounter,
-      'characterEmotion': _characterEmotion,
-      'emotionIntensity': _emotionIntensity,
-      'timeOfDay': _timeService.timeOfDay,
-      'dayCount': _timeService.dayCount,
-      'startDayOfWeek': _timeService.startDayOfWeekAnchor,
-      'arousalLevel': _nsfwService.arousalLevel,
-      'cooldownTurnsRemaining': _nsfwService.cooldownTurnsRemaining,
-      'cooldownTurnsTotal': _nsfwService.cooldownTurnsTotal,
-      'trustLevel': _relationshipService.trustLevel,
-      'activeFixation': _relationshipService.activeFixation,
-      'fixationLifespan': _relationshipService.fixationLifespan,
-      'spatialStance': _relationshipService.spatialStance,
-    };
-
-    // Include needs snapshot when the simulation is active (clean port).
-    // Note: 'enabled' is deliberately omitted from the per-message snapshot.
-    // The enabled flag is authoritative from the character card / current session
-    // (see setNeedsSimEnabled and ext seeding). Snapshots only carry the vector
-    // for timeline continuity while the sim is on. This prevents historical
-    // snapshots from resurrecting a stale enabled state after a mid-chat toggle-off.
-    if (_needsSimEnabled && _needsSimulation.vector.isNotEmpty) {
-      // Explicit <String, dynamic> for the needs snapshot so that 'deltas' (Map with
-      // mixed int/String values from computeNeedsDeltasWithReasons) can be attached
-      // without runtime generic value-type violation (the 'vector' entry statically
-      // infers Map<String,int>, which would lock the literal's value type and reject
-      // the deltas map on []=).
-      final needsSnap = <String, dynamic>{
-        'vector': Map<String, int>.from(_needsSimulation.vector),
-      };
-      state['needs'] = needsSnap;
-
-      final needsDeltas = _needsSimulation.computeNeedsDeltasWithReasons(
-        preTurn ?? const <String, int>{},
-      );
-      if (needsDeltas.isNotEmpty) {
-        needsSnap['deltas'] = needsDeltas;
-      }
-    }
-
-    return state;
-  }
-
-  // ── Phase 1: Per-character realism evaluation for the upcoming speaker ────
-  /// Fire the four realism eval calls (relationship / emotional / physical /
-  /// narrative) concurrently with the standard dispatch stagger. This exact
-  /// 4-call block was duplicated byte-for-byte in the centralized 1:1 path and
-  /// the per-speaker group path; sharing it is the first DRY step toward a single
-  /// eval path. The caller decides whether to wrap it in batched verification
-  /// (`beginCollect`/`finalize`) — that wrapping currently differs between the
-  /// two paths and is deliberately left to the caller until that divergence is
-  /// reconciled.
-  Future<void> _fireStaggeredRealismEvals(void Function(String) onChunk) async {
-    await Future.wait([
-      _evaluateRelationshipCall(onChunk: onChunk),
-      Future.delayed(
-        _kEvalDispatchStagger,
-        () => _evaluateEmotionalStateCall(onChunk: onChunk),
-      ),
-      Future.delayed(
-        _kEvalDispatchStagger * 2,
-        () => _evaluatePhysicalStateCall(onChunk: onChunk),
-      ),
-      Future.delayed(
-        _kEvalDispatchStagger * 3,
-        () => _evaluateNarrativeCall(onChunk: onChunk),
-      ),
-    ]);
-  }
 
   /// Loads the active objectives for the given character in the current session.
   /// Safe to call from group objective UIs — does not mutate global _activeObjectives.
